Connect a GitHub Repo with AWS
As part of my 7-Day DevOps Challenge, I set out to connect a GitHub repository to my Java web application hosted on an AWS EC2 instance. This was Project II in the series, and by the end of it, I had my code securely stored and version-controlled using Git and GitHub.